Fix three type conversion warnings
- Add explicit NSUInteger casts in PBGitSidebarController.m:365,367 for NSInteger to NSUInteger conversion when accessing array elements
- Add explicit NSUInteger cast in PBGitRevisionCell.m:92 for int to NSUInteger conversion in color array indexing

These fixes resolve implicit sign conversion warnings while maintaining original logic.

+++ Classes/Controllers/PBGitSidebarController.m
@@ -362,9 +362,9 @@
 - (id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)index ofItem:(id)item
 {
 	if (!item)
-		return [items objectAtIndex:index];
+		return [items objectAtIndex:(NSUInteger)index];
 
-	return [[(PBSourceViewItem *)item sortedChildren] objectAtIndex:index];
+	return [[(PBSourceViewItem *)item sortedChildren] objectAtIndex:(NSUInteger)index];
 }
 
 - (BOOL)outlineView:(NSOutlineView *)outlineView isItemExpandable:(id)item
